chore: move management services into proper packages
