+++ src/main/java/io/thinkit/edc/client/connector/services/management/Assets.java
@@ -1,60 +1,63 @@
-package io.thinkit.edc.client.connector.services;
+package io.thinkit.edc.client.connector.services.management;
 
 import static io.thinkit.edc.client.connector.utils.Constants.ID;
 import static io.thinkit.edc.client.connector.utils.JsonLdUtil.compact;
 import static java.net.http.HttpRequest.BodyPublishers.ofString;
 
 import io.thinkit.edc.client.connector.EdcClientContext;
+import io.thinkit.edc.client.connector.model.Asset;
+import io.thinkit.edc.client.connector.model.QuerySpec;
 import io.thinkit.edc.client.connector.model.Result;
-import io.thinkit.edc.client.connector.model.Secret;
 import io.thinkit.edc.client.connector.resource.management.ManagementResource;
 import io.thinkit.edc.client.connector.utils.JsonLdUtil;
 import jakarta.json.JsonArray;
 import java.net.URI;
 import java.net.http.HttpRequest;
+import java.util.List;
 import java.util.concurrent.CompletableFuture;
 
-public class Secrets extends ManagementResource {
+public class Assets extends ManagementResource {
     private final String url;
 
-    public Secrets(EdcClientContext context) {
+    public Assets(EdcClientContext context) {
         super(context);
-        url = "%s/v3/secrets".formatted(managementUrl);
+        url = "%s/v3/assets".formatted(managementUrl);
     }
 
-    public Result<Secret> get(String id) {
+    public Result<Asset> get(String id) {
         var requestBuilder = getRequestBuilder(id);
-        return context.httpClient().send(requestBuilder).map(JsonLdUtil::expand).map(this::getSecret);
+
+        return context.httpClient().send(requestBuilder).map(JsonLdUtil::expand).map(this::getAsset);
     }
 
-    public CompletableFuture<Result<Secret>> getAsync(String id) {
+    public CompletableFuture<Result<Asset>> getAsync(String id) {
         var requestBuilder = getRequestBuilder(id);
 
         return context.httpClient().sendAsync(requestBuilder).thenApply(result -> result.map(JsonLdUtil::expand)
-                .map(this::getSecret));
+                .map(this::getAsset));
     }
 
-    public Result<String> create(Secret input) {
+    public Result<String> create(Asset input) {
         var requestBuilder = createRequestBuilder(input);
 
         return context.httpClient().send(requestBuilder).map(JsonLdUtil::expand).map(content -> content.getJsonObject(0)
                 .getString(ID));
     }
 
-    public CompletableFuture<Result<String>> createAsync(Secret input) {
+    public CompletableFuture<Result<String>> createAsync(Asset input) {
         var requestBuilder = createRequestBuilder(input);
 
         return context.httpClient().sendAsync(requestBuilder).thenApply(result -> result.map(JsonLdUtil::expand)
                 .map(content -> content.getJsonObject(0).getString(ID)));
     }
 
-    public Result<String> update(Secret input) {
+    public Result<String> update(Asset input) {
         var requestBuilder = updateRequestBuilder(input);
 
         return context.httpClient().send(requestBuilder).map(result -> input.id());
     }
 
-    public CompletableFuture<Result<String>> updateAsync(Secret input) {
+    public CompletableFuture<Result<String>> updateAsync(Asset input) {
 
         var requestBuilder = updateRequestBuilder(input);
 
@@ -73,13 +76,27 @@ public class Secrets extends ManagementResource {
         return context.httpClient().sendAsync(requestBuilder).thenApply(result -> result.map(content -> id));
     }
 
+    public Result<List<Asset>> request(QuerySpec input) {
+
+        var requestBuilder = getAssetsRequestBuilder(input);
+
+        return context.httpClient().send(requestBuilder).map(JsonLdUtil::expand).map(this::getAssets);
+    }
+
+    public CompletableFuture<Result<List<Asset>>> requestAsync(QuerySpec input) {
+        var requestBuilder = getAssetsRequestBuilder(input);
+
+        return context.httpClient().sendAsync(requestBuilder).thenApply(result -> result.map(JsonLdUtil::expand)
+                .map(this::getAssets));
+    }
+
     private HttpRequest.Builder getRequestBuilder(String id) {
         return HttpRequest.newBuilder()
                 .uri(URI.create("%s/%s".formatted(this.url, id)))
                 .GET();
     }
 
-    private HttpRequest.Builder createRequestBuilder(Secret input) {
+    private HttpRequest.Builder createRequestBuilder(Asset input) {
         var requestBody = compact(input);
         return HttpRequest.newBuilder()
                 .uri(URI.create(this.url))
@@ -87,7 +104,7 @@ public class Secrets extends ManagementResource {
                 .POST(ofString(requestBody.toString()));
     }
 
-    private HttpRequest.Builder updateRequestBuilder(Secret input) {
+    private HttpRequest.Builder updateRequestBuilder(Asset input) {
         var requestBody = compact(input);
         return HttpRequest.newBuilder()
                 .uri(URI.create(this.url))
@@ -101,7 +118,21 @@ public class Secrets extends ManagementResource {
                 .DELETE();
     }
 
-    private Secret getSecret(JsonArray array) {
-        return Secret.Builder.newInstance().raw(array.getJsonObject(0)).build();
+    private HttpRequest.Builder getAssetsRequestBuilder(QuerySpec input) {
+        var requestBody = compact(input);
+        return HttpRequest.newBuilder()
+                .uri(URI.create("%s/request".formatted(this.url)))
+                .header("content-type", "application/json")
+                .POST(ofString(requestBody.toString()));
+    }
+
+    private Asset getAsset(JsonArray array) {
+        return Asset.Builder.newInstance().raw(array.getJsonObject(0)).build();
+    }
+
+    private List<Asset> getAssets(JsonArray array) {
+        return array.stream()
+                .map(s -> Asset.Builder.newInstance().raw(s.asJsonObject()).build())
+                .toList();
     }
 }
