+++ src/main/java/io/thinkit/edc/client/connector/services/management/ContractDefinitions.java
@@ -1,11 +1,11 @@
-package io.thinkit.edc.client.connector.services;
+package io.thinkit.edc.client.connector.services.management;
 
 import static io.thinkit.edc.client.connector.utils.Constants.ID;
 import static io.thinkit.edc.client.connector.utils.JsonLdUtil.compact;
 import static java.net.http.HttpRequest.BodyPublishers.ofString;
 
 import io.thinkit.edc.client.connector.EdcClientContext;
-import io.thinkit.edc.client.connector.model.Asset;
+import io.thinkit.edc.client.connector.model.ContractDefinition;
 import io.thinkit.edc.client.connector.model.QuerySpec;
 import io.thinkit.edc.client.connector.model.Result;
 import io.thinkit.edc.client.connector.resource.management.ManagementResource;
@@ -16,78 +16,79 @@ import java.net.http.HttpRequest;
 import java.util.List;
 import java.util.concurrent.CompletableFuture;
 
-public class Assets extends ManagementResource {
+public class ContractDefinitions extends ManagementResource {
     private final String url;
 
-    public Assets(EdcClientContext context) {
+    public ContractDefinitions(EdcClientContext context) {
         super(context);
-        url = "%s/v3/assets".formatted(managementUrl);
+        url = "%s/v3/contractdefinitions".formatted(managementUrl);
     }
 
-    public Result<Asset> get(String id) {
+    public Result<ContractDefinition> get(String id) {
         var requestBuilder = getRequestBuilder(id);
-
-        return context.httpClient().send(requestBuilder).map(JsonLdUtil::expand).map(this::getAsset);
+        return context.httpClient().send(requestBuilder).map(JsonLdUtil::expand).map(this::getContractDefinition);
     }
 
-    public CompletableFuture<Result<Asset>> getAsync(String id) {
+    public CompletableFuture<Result<ContractDefinition>> getAsync(String id) {
         var requestBuilder = getRequestBuilder(id);
 
         return context.httpClient().sendAsync(requestBuilder).thenApply(result -> result.map(JsonLdUtil::expand)
-                .map(this::getAsset));
+                .map(this::getContractDefinition));
     }
 
-    public Result<String> create(Asset input) {
+    public Result<String> create(ContractDefinition input) {
+
         var requestBuilder = createRequestBuilder(input);
 
         return context.httpClient().send(requestBuilder).map(JsonLdUtil::expand).map(content -> content.getJsonObject(0)
                 .getString(ID));
     }
 
-    public CompletableFuture<Result<String>> createAsync(Asset input) {
+    public CompletableFuture<Result<String>> createAsync(ContractDefinition input) {
+
         var requestBuilder = createRequestBuilder(input);
 
         return context.httpClient().sendAsync(requestBuilder).thenApply(result -> result.map(JsonLdUtil::expand)
                 .map(content -> content.getJsonObject(0).getString(ID)));
     }
 
-    public Result<String> update(Asset input) {
-        var requestBuilder = updateRequestBuilder(input);
-
-        return context.httpClient().send(requestBuilder).map(result -> input.id());
-    }
-
-    public CompletableFuture<Result<String>> updateAsync(Asset input) {
-
-        var requestBuilder = updateRequestBuilder(input);
-
-        return context.httpClient().sendAsync(requestBuilder).thenApply(result -> result.map(content -> input.id()));
-    }
-
     public Result<String> delete(String id) {
         var requestBuilder = deleteRequestBuilder(id);
-
         return context.httpClient().send(requestBuilder).map(result -> id);
     }
 
     public CompletableFuture<Result<String>> deleteAsync(String id) {
         var requestBuilder = deleteRequestBuilder(id);
-
         return context.httpClient().sendAsync(requestBuilder).thenApply(result -> result.map(content -> id));
     }
 
-    public Result<List<Asset>> request(QuerySpec input) {
+    public Result<List<ContractDefinition>> request(QuerySpec input) {
 
-        var requestBuilder = getAssetsRequestBuilder(input);
+        var requestBuilder = getContractDefinitionsRequestBuilder(input);
 
-        return context.httpClient().send(requestBuilder).map(JsonLdUtil::expand).map(this::getAssets);
+        return context.httpClient().send(requestBuilder).map(JsonLdUtil::expand).map(this::getContractDefinitions);
     }
 
-    public CompletableFuture<Result<List<Asset>>> requestAsync(QuerySpec input) {
-        var requestBuilder = getAssetsRequestBuilder(input);
+    public CompletableFuture<Result<List<ContractDefinition>>> requestAsync(QuerySpec input) {
+
+        var requestBuilder = getContractDefinitionsRequestBuilder(input);
 
         return context.httpClient().sendAsync(requestBuilder).thenApply(result -> result.map(JsonLdUtil::expand)
-                .map(this::getAssets));
+                .map(this::getContractDefinitions));
+    }
+
+    public Result<String> update(ContractDefinition input) {
+
+        var requestBuilder = updateRequestBuilder(input);
+
+        return context.httpClient().send(requestBuilder).map(result -> input.id());
+    }
+
+    public CompletableFuture<Result<String>> updateAsync(ContractDefinition input) {
+
+        var requestBuilder = updateRequestBuilder(input);
+
+        return context.httpClient().sendAsync(requestBuilder).thenApply(result -> result.map(content -> input.id()));
     }
 
     private HttpRequest.Builder getRequestBuilder(String id) {
@@ -96,7 +97,7 @@ public class Assets extends ManagementResource {
                 .GET();
     }
 
-    private HttpRequest.Builder createRequestBuilder(Asset input) {
+    private HttpRequest.Builder createRequestBuilder(ContractDefinition input) {
         var requestBody = compact(input);
         return HttpRequest.newBuilder()
                 .uri(URI.create(this.url))
@@ -104,7 +105,7 @@ public class Assets extends ManagementResource {
                 .POST(ofString(requestBody.toString()));
     }
 
-    private HttpRequest.Builder updateRequestBuilder(Asset input) {
+    private HttpRequest.Builder updateRequestBuilder(ContractDefinition input) {
         var requestBody = compact(input);
         return HttpRequest.newBuilder()
                 .uri(URI.create(this.url))
@@ -118,7 +119,7 @@ public class Assets extends ManagementResource {
                 .DELETE();
     }
 
-    private HttpRequest.Builder getAssetsRequestBuilder(QuerySpec input) {
+    private HttpRequest.Builder getContractDefinitionsRequestBuilder(QuerySpec input) {
         var requestBody = compact(input);
         return HttpRequest.newBuilder()
                 .uri(URI.create("%s/request".formatted(this.url)))
@@ -126,13 +127,17 @@ public class Assets extends ManagementResource {
                 .POST(ofString(requestBody.toString()));
     }
 
-    private Asset getAsset(JsonArray array) {
-        return Asset.Builder.newInstance().raw(array.getJsonObject(0)).build();
+    private ContractDefinition getContractDefinition(JsonArray array) {
+        return ContractDefinition.Builder.newInstance()
+                .raw(array.getJsonObject(0))
+                .build();
     }
 
-    private List<Asset> getAssets(JsonArray array) {
+    private List<ContractDefinition> getContractDefinitions(JsonArray array) {
         return array.stream()
-                .map(s -> Asset.Builder.newInstance().raw(s.asJsonObject()).build())
+                .map(s -> ContractDefinition.Builder.newInstance()
+                        .raw(s.asJsonObject())
+                        .build())
                 .toList();
     }
 }
