+++ src/main/java/io/thinkit/edc/client/connector/services/management/PolicyDefinitions.java
@@ -1,11 +1,11 @@
-package io.thinkit.edc.client.connector.services;
+package io.thinkit.edc.client.connector.services.management;
 
 import static io.thinkit.edc.client.connector.utils.Constants.ID;
 import static io.thinkit.edc.client.connector.utils.JsonLdUtil.compact;
 import static java.net.http.HttpRequest.BodyPublishers.ofString;
 
 import io.thinkit.edc.client.connector.EdcClientContext;
-import io.thinkit.edc.client.connector.model.ContractDefinition;
+import io.thinkit.edc.client.connector.model.PolicyDefinition;
 import io.thinkit.edc.client.connector.model.QuerySpec;
 import io.thinkit.edc.client.connector.model.Result;
 import io.thinkit.edc.client.connector.resource.management.ManagementResource;
@@ -16,27 +16,28 @@ import java.net.http.HttpRequest;
 import java.util.List;
 import java.util.concurrent.CompletableFuture;
 
-public class ContractDefinitions extends ManagementResource {
+public class PolicyDefinitions extends ManagementResource {
+
     private final String url;
 
-    public ContractDefinitions(EdcClientContext context) {
+    public PolicyDefinitions(EdcClientContext context) {
         super(context);
-        url = "%s/v3/contractdefinitions".formatted(managementUrl);
+        url = "%s/v3/policydefinitions".formatted(managementUrl);
     }
 
-    public Result<ContractDefinition> get(String id) {
+    public Result<PolicyDefinition> get(String id) {
         var requestBuilder = getRequestBuilder(id);
-        return context.httpClient().send(requestBuilder).map(JsonLdUtil::expand).map(this::getContractDefinition);
+
+        return context.httpClient().send(requestBuilder).map(JsonLdUtil::expand).map(this::getPolicyDefinition);
     }
 
-    public CompletableFuture<Result<ContractDefinition>> getAsync(String id) {
+    public CompletableFuture<Result<PolicyDefinition>> getAsync(String id) {
         var requestBuilder = getRequestBuilder(id);
-
         return context.httpClient().sendAsync(requestBuilder).thenApply(result -> result.map(JsonLdUtil::expand)
-                .map(this::getContractDefinition));
+                .map(this::getPolicyDefinition));
     }
 
-    public Result<String> create(ContractDefinition input) {
+    public Result<String> create(PolicyDefinition input) {
 
         var requestBuilder = createRequestBuilder(input);
 
@@ -44,7 +45,7 @@ public class ContractDefinitions extends ManagementResource {
                 .getString(ID));
     }
 
-    public CompletableFuture<Result<String>> createAsync(ContractDefinition input) {
+    public CompletableFuture<Result<String>> createAsync(PolicyDefinition input) {
 
         var requestBuilder = createRequestBuilder(input);
 
@@ -52,43 +53,45 @@ public class ContractDefinitions extends ManagementResource {
                 .map(content -> content.getJsonObject(0).getString(ID)));
     }
 
-    public Result<String> delete(String id) {
-        var requestBuilder = deleteRequestBuilder(id);
-        return context.httpClient().send(requestBuilder).map(result -> id);
-    }
-
-    public CompletableFuture<Result<String>> deleteAsync(String id) {
-        var requestBuilder = deleteRequestBuilder(id);
-        return context.httpClient().sendAsync(requestBuilder).thenApply(result -> result.map(content -> id));
-    }
-
-    public Result<List<ContractDefinition>> request(QuerySpec input) {
-
-        var requestBuilder = getContractDefinitionsRequestBuilder(input);
-
-        return context.httpClient().send(requestBuilder).map(JsonLdUtil::expand).map(this::getContractDefinitions);
-    }
-
-    public CompletableFuture<Result<List<ContractDefinition>>> requestAsync(QuerySpec input) {
-
-        var requestBuilder = getContractDefinitionsRequestBuilder(input);
-
-        return context.httpClient().sendAsync(requestBuilder).thenApply(result -> result.map(JsonLdUtil::expand)
-                .map(this::getContractDefinitions));
-    }
-
-    public Result<String> update(ContractDefinition input) {
+    public Result<String> update(PolicyDefinition input) {
 
         var requestBuilder = updateRequestBuilder(input);
 
         return context.httpClient().send(requestBuilder).map(result -> input.id());
     }
 
-    public CompletableFuture<Result<String>> updateAsync(ContractDefinition input) {
+    public CompletableFuture<Result<String>> updateAsync(PolicyDefinition input) {
 
         var requestBuilder = updateRequestBuilder(input);
 
         return context.httpClient().sendAsync(requestBuilder).thenApply(result -> result.map(content -> input.id()));
+    }
+
+    public Result<String> delete(String id) {
+        var requestBuilder = deleteRequestBuilder(id);
+
+        return context.httpClient().send(requestBuilder).map(result -> id);
+    }
+
+    public CompletableFuture<Result<String>> deleteAsync(String id) {
+        var requestBuilder = deleteRequestBuilder(id);
+
+        return context.httpClient().sendAsync(requestBuilder).thenApply(result -> result.map(content -> id));
+    }
+
+    public Result<List<PolicyDefinition>> request(QuerySpec input) {
+
+        var requestBuilder = getPolicyDefinitionsRequestBuilder(input);
+
+        return context.httpClient().send(requestBuilder).map(JsonLdUtil::expand).map(this::getPolicyDefinitions);
+    }
+
+    public CompletableFuture<Result<List<PolicyDefinition>>> requestAsync(QuerySpec input) {
+
+        var requestBuilder = getPolicyDefinitionsRequestBuilder(input);
+
+        return context.httpClient().sendAsync(requestBuilder).thenApply(result -> result.map(JsonLdUtil::expand)
+                .map(this::getPolicyDefinitions));
     }
 
     private HttpRequest.Builder getRequestBuilder(String id) {
@@ -97,7 +100,7 @@ public class ContractDefinitions extends ManagementResource {
                 .GET();
     }
 
-    private HttpRequest.Builder createRequestBuilder(ContractDefinition input) {
+    private HttpRequest.Builder createRequestBuilder(PolicyDefinition input) {
         var requestBody = compact(input);
         return HttpRequest.newBuilder()
                 .uri(URI.create(this.url))
@@ -105,10 +108,10 @@ public class ContractDefinitions extends ManagementResource {
                 .POST(ofString(requestBody.toString()));
     }
 
-    private HttpRequest.Builder updateRequestBuilder(ContractDefinition input) {
+    private HttpRequest.Builder updateRequestBuilder(PolicyDefinition input) {
         var requestBody = compact(input);
         return HttpRequest.newBuilder()
-                .uri(URI.create(this.url))
+                .uri(URI.create("%s/%s".formatted(this.url, input.id())))
                 .header("content-type", "application/json")
                 .PUT(ofString(requestBody.toString()));
     }
@@ -119,7 +122,7 @@ public class ContractDefinitions extends ManagementResource {
                 .DELETE();
     }
 
-    private HttpRequest.Builder getContractDefinitionsRequestBuilder(QuerySpec input) {
+    private HttpRequest.Builder getPolicyDefinitionsRequestBuilder(QuerySpec input) {
         var requestBody = compact(input);
         return HttpRequest.newBuilder()
                 .uri(URI.create("%s/request".formatted(this.url)))
@@ -127,15 +130,15 @@ public class ContractDefinitions extends ManagementResource {
                 .POST(ofString(requestBody.toString()));
     }
 
-    private ContractDefinition getContractDefinition(JsonArray array) {
-        return ContractDefinition.Builder.newInstance()
+    private PolicyDefinition getPolicyDefinition(JsonArray array) {
+        return PolicyDefinition.Builder.newInstance()
                 .raw(array.getJsonObject(0))
                 .build();
     }
 
-    private List<ContractDefinition> getContractDefinitions(JsonArray array) {
+    private List<PolicyDefinition> getPolicyDefinitions(JsonArray array) {
         return array.stream()
-                .map(s -> ContractDefinition.Builder.newInstance()
+                .map(s -> PolicyDefinition.Builder.newInstance()
                         .raw(s.asJsonObject())
                         .build())
                 .toList();
