+++ src/main/java/io/thinkit/edc/client/connector/services/management/Secrets.java
@@ -1,66 +1,60 @@
-package io.thinkit.edc.client.connector.services;
+package io.thinkit.edc.client.connector.services.management;
 
 import static io.thinkit.edc.client.connector.utils.Constants.ID;
 import static io.thinkit.edc.client.connector.utils.JsonLdUtil.compact;
 import static java.net.http.HttpRequest.BodyPublishers.ofString;
 
 import io.thinkit.edc.client.connector.EdcClientContext;
-import io.thinkit.edc.client.connector.model.PolicyDefinition;
-import io.thinkit.edc.client.connector.model.QuerySpec;
 import io.thinkit.edc.client.connector.model.Result;
+import io.thinkit.edc.client.connector.model.Secret;
 import io.thinkit.edc.client.connector.resource.management.ManagementResource;
 import io.thinkit.edc.client.connector.utils.JsonLdUtil;
 import jakarta.json.JsonArray;
 import java.net.URI;
 import java.net.http.HttpRequest;
-import java.util.List;
 import java.util.concurrent.CompletableFuture;
 
-public class PolicyDefinitions extends ManagementResource {
-
+public class Secrets extends ManagementResource {
     private final String url;
 
-    public PolicyDefinitions(EdcClientContext context) {
+    public Secrets(EdcClientContext context) {
         super(context);
-        url = "%s/v3/policydefinitions".formatted(managementUrl);
+        url = "%s/v3/secrets".formatted(managementUrl);
     }
 
-    public Result<PolicyDefinition> get(String id) {
+    public Result<Secret> get(String id) {
         var requestBuilder = getRequestBuilder(id);
-
-        return context.httpClient().send(requestBuilder).map(JsonLdUtil::expand).map(this::getPolicyDefinition);
+        return context.httpClient().send(requestBuilder).map(JsonLdUtil::expand).map(this::getSecret);
     }
 
-    public CompletableFuture<Result<PolicyDefinition>> getAsync(String id) {
+    public CompletableFuture<Result<Secret>> getAsync(String id) {
         var requestBuilder = getRequestBuilder(id);
+
         return context.httpClient().sendAsync(requestBuilder).thenApply(result -> result.map(JsonLdUtil::expand)
-                .map(this::getPolicyDefinition));
+                .map(this::getSecret));
     }
 
-    public Result<String> create(PolicyDefinition input) {
-
+    public Result<String> create(Secret input) {
         var requestBuilder = createRequestBuilder(input);
 
         return context.httpClient().send(requestBuilder).map(JsonLdUtil::expand).map(content -> content.getJsonObject(0)
                 .getString(ID));
     }
 
-    public CompletableFuture<Result<String>> createAsync(PolicyDefinition input) {
-
+    public CompletableFuture<Result<String>> createAsync(Secret input) {
         var requestBuilder = createRequestBuilder(input);
 
         return context.httpClient().sendAsync(requestBuilder).thenApply(result -> result.map(JsonLdUtil::expand)
                 .map(content -> content.getJsonObject(0).getString(ID)));
     }
 
-    public Result<String> update(PolicyDefinition input) {
-
+    public Result<String> update(Secret input) {
         var requestBuilder = updateRequestBuilder(input);
 
         return context.httpClient().send(requestBuilder).map(result -> input.id());
     }
 
-    public CompletableFuture<Result<String>> updateAsync(PolicyDefinition input) {
+    public CompletableFuture<Result<String>> updateAsync(Secret input) {
 
         var requestBuilder = updateRequestBuilder(input);
 
@@ -79,28 +73,13 @@ public class PolicyDefinitions extends ManagementResource {
         return context.httpClient().sendAsync(requestBuilder).thenApply(result -> result.map(content -> id));
     }
 
-    public Result<List<PolicyDefinition>> request(QuerySpec input) {
-
-        var requestBuilder = getPolicyDefinitionsRequestBuilder(input);
-
-        return context.httpClient().send(requestBuilder).map(JsonLdUtil::expand).map(this::getPolicyDefinitions);
-    }
-
-    public CompletableFuture<Result<List<PolicyDefinition>>> requestAsync(QuerySpec input) {
-
-        var requestBuilder = getPolicyDefinitionsRequestBuilder(input);
-
-        return context.httpClient().sendAsync(requestBuilder).thenApply(result -> result.map(JsonLdUtil::expand)
-                .map(this::getPolicyDefinitions));
-    }
-
     private HttpRequest.Builder getRequestBuilder(String id) {
         return HttpRequest.newBuilder()
                 .uri(URI.create("%s/%s".formatted(this.url, id)))
                 .GET();
     }
 
-    private HttpRequest.Builder createRequestBuilder(PolicyDefinition input) {
+    private HttpRequest.Builder createRequestBuilder(Secret input) {
         var requestBody = compact(input);
         return HttpRequest.newBuilder()
                 .uri(URI.create(this.url))
@@ -108,10 +87,10 @@ public class PolicyDefinitions extends ManagementResource {
                 .POST(ofString(requestBody.toString()));
     }
 
-    private HttpRequest.Builder updateRequestBuilder(PolicyDefinition input) {
+    private HttpRequest.Builder updateRequestBuilder(Secret input) {
         var requestBody = compact(input);
         return HttpRequest.newBuilder()
-                .uri(URI.create("%s/%s".formatted(this.url, input.id())))
+                .uri(URI.create(this.url))
                 .header("content-type", "application/json")
                 .PUT(ofString(requestBody.toString()));
     }
@@ -122,25 +101,7 @@ public class PolicyDefinitions extends ManagementResource {
                 .DELETE();
     }
 
-    private HttpRequest.Builder getPolicyDefinitionsRequestBuilder(QuerySpec input) {
-        var requestBody = compact(input);
-        return HttpRequest.newBuilder()
-                .uri(URI.create("%s/request".formatted(this.url)))
-                .header("content-type", "application/json")
-                .POST(ofString(requestBody.toString()));
-    }
-
-    private PolicyDefinition getPolicyDefinition(JsonArray array) {
-        return PolicyDefinition.Builder.newInstance()
-                .raw(array.getJsonObject(0))
-                .build();
-    }
-
-    private List<PolicyDefinition> getPolicyDefinitions(JsonArray array) {
-        return array.stream()
-                .map(s -> PolicyDefinition.Builder.newInstance()
-                        .raw(s.asJsonObject())
-                        .build())
-                .toList();
+    private Secret getSecret(JsonArray array) {
+        return Secret.Builder.newInstance().raw(array.getJsonObject(0)).build();
     }
 }
