+++ src/test/java/io/thinkit/edc/client/connector/services/management/SecretsTest.java
@@ -1,4 +1,4 @@
-package io.thinkit.edc.client.connector.services;
+package io.thinkit.edc.client.connector.services.management;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -8,6 +8,7 @@ import io.thinkit.edc.client.connector.model.Result;
 import io.thinkit.edc.client.connector.model.Secret;
 import java.net.http.HttpClient;
 import java.util.concurrent.TimeUnit;
+
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
